Derive header user in auth proxy middleware
